smartcare_django/requirements.txt
Django==5.0.2
mysql-connector-python==8.3.0
django-mysql
djangorestframework
django-rest-knox